HomeView refactored
