+++ VultisigApp/VultisigApp/Views/Vault/HomeView.swift
@@ -30,78 +30,7 @@ struct HomeView: View {
     @Environment(\.modelContext) private var modelContext
     
     var body: some View {
-        ZStack {
-            Background()
-            main
-        }
-#if os(macOS)
-        .alert(
-            NSLocalizedString("newUpdateAvailable", comment: ""),
-            isPresented: $checkUpdateViewModel.showUpdateAlert
-        ) {
-            Link(destination: URL(string: Endpoint.githubMacUpdateBase + checkUpdateViewModel.latestVersionBase)!) {
-                Text(NSLocalizedString("updateNow", comment: ""))
-            }
-            
-            Button(NSLocalizedString("dismiss", comment: ""), role: .cancel) {}
-        } message: {
-            Text(checkUpdateViewModel.latestVersion)
-        }
-#endif
-    }
-    
-    var main: some View {
-        VStack(spacing: 0) {
-#if os(macOS)
-            headerMac
-            Separator()
-#endif
-            view
-        }
-    }
-    
-    var headerMac: some View {
-        HomeHeader(
-            showVaultsList: $showVaultsList,
-            isEditingVaults: $isEditingVaults
-        )
-    }
-    
-    var view: some View {
-        VStack(spacing: 0) {
-            ZStack {
-                if let vault = viewModel.selectedVault {
-                    VaultDetailView(showVaultsList: $showVaultsList, vault: vault)
-                }
-                
-                VaultsView(viewModel: viewModel, showVaultsList: $showVaultsList, isEditingVaults: $isEditingVaults)
-            }
-        }
-        .navigationBarBackButtonHidden(true)
-#if os(iOS)
-        .toolbar {
-            ToolbarItem(placement: Placement.topBarLeading.getPlacement()) {
-                menuButton
-            }
-            ToolbarItem(placement: Placement.principal.getPlacement()) {
-                navigationTitle
-            }
-            ToolbarItem(placement: Placement.topBarTrailing.getPlacement()) {
-                editButton
-            }
-        }
-#endif
-        .onAppear {
-            setData()
-        }
-        .navigationDestination(isPresented: $shouldJoinKeygen) {
-            JoinKeygenView(vault: Vault(name: "Main Vault"))
-        }
-        .navigationDestination(isPresented: $shouldKeysignTransaction) {
-            if let vault = viewModel.selectedVault {
-                JoinKeysignView(vault: vault)
-            }
-        }
+        content
     }
     
     var navigationTitle: some View {
@@ -152,28 +81,7 @@ struct HomeView: View {
         )
     }
     
-    private func setData() {
-        fetchVaults()
-        shouldJoinKeygen = false
-        shouldKeysignTransaction = false
-        
-#if os(macOS)
-        macCameraServiceViewModel.stopSession()
-        checkUpdateViewModel.checkForUpdates(isAutoCheck: true)
-#endif
-        
-        if let vault = selectedVault {
-            viewModel.setSelectedVault(vault)
-            selectedVault = nil
-            return
-        } else {
-            viewModel.loadSelectedVault(for: vaults)
-        }
-        
-        presetValuesForDeeplink()
-    }
-    
-    private func presetValuesForDeeplink() {
+    func presetValuesForDeeplink() {
         guard let type = deeplinkViewModel.type else {
             return
         }
@@ -207,7 +115,7 @@ struct HomeView: View {
         }
     }
     
-    private func switchView() {
+    func switchView() {
         guard viewModel.selectedVault != nil else {
             return
         }
@@ -217,7 +125,7 @@ struct HomeView: View {
         }
     }
     
-    private func fetchVaults() {
+    func fetchVaults() {
         let fetchVaultDescriptor = FetchDescriptor<Vault>()
         do {
             vaults = try modelContext.fetch(fetchVaultDescriptor)
@@ -231,8 +139,6 @@ struct HomeView: View {
     HomeView()
         .environmentObject(DeeplinkViewModel())
         .environmentObject(HomeViewModel())
-#if os(macOS)
         .environmentObject(MacCameraServiceViewModel())
         .environmentObject(CheckUpdateViewModel())
-#endif
 }
