+++ VultisigApp/VultisigApp/iOS/View/HomeView+iOS.swift
@@ -1,0 +1,74 @@
+//
+//  HomeView+iOS.swift
+//  VultisigApp
+//
+//  Created by Amol Kumar on 2024-09-19.
+//
+
+#if os(iOS)
+import SwiftUI
+
+extension HomeView {
+    var content: some View {
+        ZStack {
+            Background()
+            main
+        }
+    }
+    
+    var main: some View {
+        view
+    }
+    
+    var view: some View {
+        VStack(spacing: 0) {
+            ZStack {
+                if let vault = viewModel.selectedVault {
+                    VaultDetailView(showVaultsList: $showVaultsList, vault: vault)
+                }
+                
+                VaultsView(viewModel: viewModel, showVaultsList: $showVaultsList, isEditingVaults: $isEditingVaults)
+            }
+        }
+        .navigationBarBackButtonHidden(true)
+        .toolbar {
+            ToolbarItem(placement: Placement.topBarLeading.getPlacement()) {
+                menuButton
+            }
+            ToolbarItem(placement: Placement.principal.getPlacement()) {
+                navigationTitle
+            }
+            ToolbarItem(placement: Placement.topBarTrailing.getPlacement()) {
+                editButton
+            }
+        }
+        .onAppear {
+            setData()
+        }
+        .navigationDestination(isPresented: $shouldJoinKeygen) {
+            JoinKeygenView(vault: Vault(name: "Main Vault"))
+        }
+        .navigationDestination(isPresented: $shouldKeysignTransaction) {
+            if let vault = viewModel.selectedVault {
+                JoinKeysignView(vault: vault)
+            }
+        }
+    }
+    
+    private func setData() {
+        fetchVaults()
+        shouldJoinKeygen = false
+        shouldKeysignTransaction = false
+        
+        if let vault = selectedVault {
+            viewModel.setSelectedVault(vault)
+            selectedVault = nil
+            return
+        } else {
+            viewModel.loadSelectedVault(for: vaults)
+        }
+        
+        presetValuesForDeeplink()
+    }
+}
+#endif
